+++ src/types/join.ts
@@ -1,0 +1,10 @@
+export interface joinForm {
+  username?: string;
+  userId?: string;
+  password?: string;
+  confirmPassword?: string;
+  email?: string;
+  phone?: string;
+  location?: string;
+  avatar?: string;
+}
